Extract validation logic for scheduled publishing fields into a mixin
Use the mixin for models that extend DraftStateMixin
