+++ wagtail/admin/forms/pages.py
@@ -1,6 +1,5 @@
 from django import forms
 from django.conf import settings
-from django.utils import timezone
 from django.utils.translation import gettext as _
 from django.utils.translation import ngettext
 
@@ -155,10 +154,6 @@ class WagtailAdminPageForm(WagtailAdminModelForm):
             del self.fields["comment_notifications"]
 
     @property
-    def show_schedule_publishing_toggle(self):
-        return "go_live_at" in self.__class__.base_fields
-
-    @property
     def show_comments_toggle(self):
         return "comments" in self.__class__.formsets
 
@@ -189,31 +184,6 @@ class WagtailAdminPageForm(WagtailAdminModelForm):
                 self.add_error(
                     "slug", forms.ValidationError(_("This slug is already in use"))
                 )
-
-        # Check scheduled publishing fields
-        go_live_at = cleaned_data.get("go_live_at")
-        expire_at = cleaned_data.get("expire_at")
-
-        # Go live must be before expire
-        if go_live_at and expire_at:
-            if go_live_at > expire_at:
-                msg = _("Go live date/time must be before expiry date/time")
-                self.add_error("go_live_at", forms.ValidationError(msg))
-                self.add_error("expire_at", forms.ValidationError(msg))
-
-        # Expire at must be in the future
-        if expire_at and expire_at < timezone.now():
-            self.add_error(
-                "expire_at",
-                forms.ValidationError(_("Expiry date/time must be in the future")),
-            )
-
-        # Don't allow an existing first_published_at to be unset by clearing the field
-        if (
-            "first_published_at" in cleaned_data
-            and not cleaned_data["first_published_at"]
-        ):
-            del cleaned_data["first_published_at"]
 
         return cleaned_data
